Re organizing prep cell structure
