+++ web/src/features/hfiCalculator/components/DailyViewTable.tsx
@@ -289,7 +289,6 @@ export const DailyViewTable = (props: Props): JSX.Element => {
                           testid={`daily-prep-level-${areaName}`}
                           meanIntensityGroup={meanIntensityGroup}
                           areaName={areaName}
-                          meanPrepLevelBoolean={false}
                         />
                       </TableRow>
                       {Object.entries(area.stations)
